refactor(providers): extract image factory to image_factory.py
Move create_image_provider() from image_openai.py to a dedicated
image_factory.py module with lazy imports. This prepares for adding
placeholder and A1111 providers without coupling them to the OpenAI
module.
